requirements.txt
asgiref==3.8.1
Django==5.1.4
django-formtools==2.5.1
sqlparse==0.5.2
tzdata==2024.2
